Move the evaluation step to process/provisioning
